mofidune/requirements.txt
asgiref==3.6.0
attrs==22.2.0
autopep8==2.0.1
black==22.12.0
click==8.1.3
coverage==7.0.4
Django==4.1.5
django-js-asset==2.0.0
django-mptt==0.14.0
djangorestframework==3.14.0
drf-spectacular==0.25.1
exceptiongroup==1.1.0
factory-boy==3.2.1
Faker==15.3.4
flake8==6.0.0
inflection==0.5.1
iniconfig==2.0.0
jsonschema==4.17.3
mccabe==0.7.0
mypy-extensions==0.4.3
packaging==22.0
pathspec==0.10.3
platformdirs==2.6.2
pluggy==1.0.0
pycodestyle==2.10.0
pyflakes==3.0.1
pyrsistent==0.19.3
pytest==7.2.0
pytest-cov==4.0.0
pytest-django==4.5.2
pytest-factoryboy==2.5.1
python-dateutil==2.8.2
python-dotenv==0.21.0
pytz==2022.7
PyYAML==6.0
six==1.16.0
sqlparse==0.4.3
tomli==2.0.1
typing_extensions==4.4.0
uritemplate==4.1.1
